made startup and internet safety, bug fixes

+++ LanguageCenter/src/main/java/com/novasa/languagecenter/languagecenterliabary_features/presentation/LCViewModel.kt
@@ -3,9 +3,9 @@ package com.novasa.languagecenter.languagecenterliabary_features.presentation
 import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import androidx.lifecycle.viewmodel.compose.viewModel
 import com.novasa.languagecenter.languagecenterliabary_features.data.repostory.ApiRepository
 import com.novasa.languagecenter.languagecenterliabary_features.data.repostory.DaoRepository
-import com.novasa.languagecenter.languagecenterliabary_features.domain.api_models.AccountInfoModel
 import com.novasa.languagecenter.languagecenterliabary_features.domain.api_models.LanguageModel
 import com.novasa.languagecenter.languagecenterliabary_features.domain.api_models.StringModel
 import com.novasa.languagecenter.languagecenterliabary_features.domain.dao_models.LanguageEntity
@@ -16,13 +16,13 @@ import com.novasa.languagecenter.languagecenterliabary_features.use_cases.HasInt
 import com.novasa.languagecenter.languagecenterliabary_features.use_cases.UnixConverter.getDateTime
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.*
 import kotlinx.coroutines.launch
-import kotlinx.coroutines.runBlocking
 import java.io.IOException
-import java.util.*
+import java.util.Collections.list
 import javax.inject.Inject
-import kotlin.collections.ArrayList
+import javax.inject.Provider
 
 @HiltViewModel
 class LCViewModel @Inject constructor(
@@ -39,7 +39,6 @@ class LCViewModel @Inject constructor(
         READY,
         FAILED
     }
-
     private val daoTranslationsToMap = daoRepository.getTranslations()
         .map { list -> list.associateBy { it.key } }
         .stateIn(
@@ -53,15 +52,32 @@ class LCViewModel @Inject constructor(
     val currentStatus: StateFlow<Status>
         get() = _currentStatus.asStateFlow()
 
+    val response: StateFlow<Map<String, TranslationEntity>>
+        get() = daoTranslationsToMap
+
     fun getReponse(
         category: String,
         key: String,
         value: String,
-    ): StateFlow<TranslationEntity?> {
-        val stateFlowOfTranslationEntity = MutableStateFlow(daoTranslationsToMap.value[key])
+    ): Flow<TranslationEntity?> {
+        val stateFlowOfTranslationEntity = flowOf(daoTranslationsToMap.value[key]).stateIn(
+            scope = viewModelScope,
+            started = SharingStarted.Eagerly,
+            initialValue = TranslationEntity("","","")
+        )
 
         viewModelScope.launch(Dispatchers.IO) {
-            if (daoTranslationsToMap.value[key] == null) {
+            provider.setCurrentLanguage(
+                configureLanguage.languageConfiguring(
+                    languageList = api.getListLanguage(),
+                    selectedLanguage = provider.language
+                )
+            )
+
+            val mapOfStrings = api.getListStrings(provider.currentLanguage)
+                .associateBy({it.key}, {it})
+
+            if (mapOfStrings[key] == null) {
                 if (hasInternet.hasInternet()) {
                     postString(
                         category = category,
@@ -69,28 +85,24 @@ class LCViewModel @Inject constructor(
                         value = value
                     )
                 }
-                runBlocking {
-                    val translationData = daoTranslationsToMap.value[key]
-                    daoRepository.insertTranslation(
-                        TranslationEntity(
-                            key = translationData!!.key,
-                            value = translationData.value,
-                            language = provider.currentLanguage
-                        )
+            }
+            if (daoTranslationsToMap.value[key] == null) {
+                Log.d("best flow", "hello")
+                daoRepository.insertTranslation(
+                    TranslationEntity(
+                        key = key,
+                        value = value,
+                        language = category
                     )
-                }
+                )
             }
         }
 
         return responeInstance(stateFlowOfTranslationEntity).response
     }
 
-    //removeable fun
-    fun testDeleteDao() {
-        viewModelScope.launch(Dispatchers.IO) {
-            daoRepository.deleteItem(daoRepository.getLanguageInfo())
-            Log.d("deleted", daoTranslationsToMap.value.toString())
-        }
+    fun morphtest() {
+        Log.d("best flow", daoTranslationsToMap.value.toString())
     }
 
 
@@ -99,10 +111,9 @@ class LCViewModel @Inject constructor(
         key: String,
         value: String,
     ) {
-        viewModelScope.launch {
+        viewModelScope.launch(Dispatchers.IO) {
             try {
                 api.postString(category, key, value)
-                getListStrings()
             } catch (e: IOException) {
                 Log.d("MainActivity", "$e")
             }
@@ -119,17 +130,27 @@ class LCViewModel @Inject constructor(
                 )
             )
             try {
-                val localLanguageInfo = daoRepository.getLanguageInfo()
+                val localLanguageInfo = if (daoRepository.getLanguageInfo() == null) {
+                    daoRepository.getLanguageInfo()
+                } else {
+                    LanguageEntity(
+                        "",
+                        "",
+                        false,
+                        0
+                    )
+                }
                 val remoteLanguageInfo = if (hasInternet.hasInternet()) {
                     api.getSpecificLanguage(provider.currentLanguage)
                 } else {
                     LanguageModel(
-                        name = "",
-                        codename = "",
+                        name = "danish",
+                        codename = "da",
                         is_fallback = false,
                         timestamp = 0,
                     )
                 }
+
                 val remoteTranslations = if (hasInternet.hasInternet()) {
                     api.getListStrings(provider.currentLanguage)
                 } else {
@@ -141,21 +162,22 @@ class LCViewModel @Inject constructor(
                         htmlTags = arrayOf("",""),
                     ))
                 }
-
-                Log.d("html", "${remoteTranslations}")
-
                 if (localLanguageInfo == null || getDateTime(remoteLanguageInfo.timestamp) >= localLanguageInfo.timestamp
                     && getDateTime(remoteLanguageInfo.timestamp) != localLanguageInfo.timestamp
                 ) {
                     _currentStatus.value = Status.UPDATING
-                    daoRepository.deleteItem(daoRepository.getLanguageInfo())
+
+                    if(hasInternet.hasInternet()){
+                        if (localLanguageInfo != null) {
+                            daoRepository.deleteItem(daoRepository.getLanguageInfo())
+                        }
+                    }
                     daoRepository.insertLanguage(
                         languageEntity = LanguageEntity(
                             name = remoteLanguageInfo.name,
                             codename =  remoteLanguageInfo.codename,
                             is_fallback =  remoteLanguageInfo.is_fallback,
-                            timestamp = getDateTime(remoteLanguageInfo.timestamp
-                            )
+                            timestamp = getDateTime(remoteLanguageInfo.timestamp)
                         )
                     )
                     daoRepository.insertTranslations(
@@ -176,11 +198,12 @@ class LCViewModel @Inject constructor(
                 _currentStatus.value = Status.READY
             }
         }
+        Log.d("hello", daoTranslationsToMap.value.toString())
     }
 }
 
 class responeInstance(
-    private val stateFlowOfTranslationEntity: MutableStateFlow<TranslationEntity?>
+    private val stateFlowOfTranslationEntity: StateFlow<TranslationEntity?>
 ) {
     val response: StateFlow<TranslationEntity?>
         get() = stateFlowOfTranslationEntity
